requirements.txt
spacy>=3.4.0
faker>=15.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.28.0
colorama>=0.4.6
pytest>=7.0.0
ttkbootstrap>=1.4.0
google-generativeai>=0.3.0
flask>=2.0.0
flask-cors>=3.0.0
